app/requirements.txt
Django==1.11.5
django-geoposition==0.3.0
geopy==1.11.0
olefile==0.44
Pillow==4.2.1
pytz==2017.2
